Agent/pyproject.toml
[project]
name = "demo-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.31.1",
    "google-genai>=1.70.0",
    "python-dotenv>=1.2.2",
    "vertexai>=1.43.0",
]
